central-backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-cors-headers==3.14.0
django-extensions==3.1.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==13.3.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
